refactor: consolidating `azuremanagedlustrefilesystem` into `hpccache` for now
This'll get renamed to `storagecache` shortly
