+++ internal/services/hpccache/managed_lustre_file_system_resource_test.go
@@ -1,4 +1,4 @@
-package azuremanagedlustrefilesystem_test
+package hpccache_test
 
 import (
 	"context"
